refactor process intrinsic
