+++ src/rust/engine/src/intrinsics/mod.rs
@@ -1,27 +1,22 @@
 // Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-use std::time::Duration;
+use futures::future::{BoxFuture, FutureExt};
+use indexmap::IndexMap;
+use rule_graph::{DependencyKey, RuleId};
 
 use crate::context::Context;
-use crate::externs;
-use crate::nodes::{ExecuteProcess, NodeResult, RunId, SessionValues, Snapshot};
+use crate::nodes::{NodeResult, RunId, SessionValues};
 use crate::python::Value;
 use crate::tasks::Intrinsic;
 use crate::types::Types;
-
-use futures::future::{BoxFuture, FutureExt, TryFutureExt};
-use futures::try_join;
-use indexmap::IndexMap;
-use pyo3::{IntoPy, Python};
-
-use rule_graph::{DependencyKey, RuleId};
 
 // Sub-modules with intrinsic implementations.
 mod dep_inference;
 mod digests;
 mod docker;
 mod interactive_process;
+mod process;
 
 use self::dep_inference::{parse_javascript_deps, parse_python_deps};
 use self::digests::{
@@ -32,6 +27,7 @@ use self::digests::{
 };
 use self::docker::docker_resolve_image;
 use self::interactive_process::interactive_process;
+use self::process::process_request_to_process_result;
 
 type IntrinsicFn =
     Box<dyn Fn(Context, Vec<Value>) -> BoxFuture<'static, NodeResult<Value>> + Send + Sync>;
@@ -204,75 +200,6 @@ impl Intrinsics {
             .unwrap_or_else(|| panic!("Unrecognized intrinsic: {intrinsic:?}"));
         function(context, args).await
     }
-}
-
-fn process_request_to_process_result(
-    context: Context,
-    mut args: Vec<Value>,
-) -> BoxFuture<'static, NodeResult<Value>> {
-    async move {
-        let process_config: externs::process::PyProcessExecutionEnvironment =
-            Python::with_gil(|py| {
-                args.pop()
-                    .unwrap()
-                    .as_ref()
-                    .extract(py)
-                    .map_err(|e| format!("{e}"))
-            })?;
-        let process_request =
-            ExecuteProcess::lift(&context.core.store(), args.pop().unwrap(), process_config)
-                .map_err(|e| e.enrich("Error lifting Process"))
-                .await?;
-
-        let result = context.get(process_request).await?.result;
-
-        let store = context.core.store();
-        let (stdout_bytes, stderr_bytes) = try_join!(
-            store
-                .load_file_bytes_with(result.stdout_digest, |bytes: &[u8]| bytes.to_owned())
-                .map_err(|e| e.enrich("Bytes from stdout")),
-            store
-                .load_file_bytes_with(result.stderr_digest, |bytes: &[u8]| bytes.to_owned())
-                .map_err(|e| e.enrich("Bytes from stderr"))
-        )?;
-
-        Python::with_gil(|py| -> NodeResult<Value> {
-            Ok(externs::unsafe_call(
-                py,
-                context.core.types.process_result,
-                &[
-                    externs::store_bytes(py, &stdout_bytes),
-                    Snapshot::store_file_digest(py, result.stdout_digest)?,
-                    externs::store_bytes(py, &stderr_bytes),
-                    Snapshot::store_file_digest(py, result.stderr_digest)?,
-                    externs::store_i64(py, result.exit_code.into()),
-                    Snapshot::store_directory_digest(py, result.output_directory)?,
-                    externs::unsafe_call(
-                        py,
-                        context.core.types.process_result_metadata,
-                        &[
-                            result
-                                .metadata
-                                .total_elapsed
-                                .map(|d| {
-                                    externs::store_u64(py, Duration::from(d).as_millis() as u64)
-                                })
-                                .unwrap_or_else(|| Value::from(py.None())),
-                            Value::from(
-                                externs::process::PyProcessExecutionEnvironment {
-                                    environment: result.metadata.environment,
-                                }
-                                .into_py(py),
-                            ),
-                            externs::store_utf8(py, result.metadata.source.into()),
-                            externs::store_u64(py, result.metadata.source_run_id.0.into()),
-                        ],
-                    ),
-                ],
-            ))
-        })
-    }
-    .boxed()
 }
 
 fn session_values(context: Context, _args: Vec<Value>) -> BoxFuture<'static, NodeResult<Value>> {
